// src/app/api/model/timeslot-preset-request.ts
/**
 * Conference Schedule Optimizer API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TimeslotPresetRequest { 
    room: string;
    /**
     * Start time of the timeslot (HH:mm)
     */
    start?: string;
    /**
     * End time of the timeslot (HH:mm)
     */
    end?: string;
    seats: number;
    beamerAvailable?: boolean;
    flipchartAvailable?: boolean;
}