// src/app/api/model/timeslot-preset-view-model.ts
/**
 * Conference Schedule Optimizer API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TimeslotPresetViewModel { 
    id: number;
    room: string;
    /**
     * Start time of the timeslot (HH:mm)
     */
    start?: string;
    /**
     * End time of the timeslot (HH:mm)
     */
    end?: string;
    seats: number;
    beamerAvailable?: boolean;
    flipchartAvailable?: boolean;
    /**
     * Creation timestamp (ISO-8601), read-only
     */
    created?: string;
    /**
     * Last modification timestamp (ISO-8601), read-only
     */
    updated?: string;
    /**
     * User who created the preset, read-only
     */
    createdBy?: string;
    /**
     * User who last modified the preset, read-only
     */
    updatedBy?: string;
}